Refactor Inputs on Login into its own Component to fix playback error regarding Demo/Regular Login
